requirements.txt
flask[async]
aiohttp
SQLAlchemy
babel
mysqlclient
waitress
nats-py
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-aiohttp-client
